Created index.html
This commit creates a HTML page called index.html with the correct meta-data for this task.
The HTML file's body contains a message to the user to direct them to check their console log and it executes the javascript file named script.js which is currently empty so it does not output anything onto the console.