Find Specific by Title and Author
This commit searched all the pages in linear order until the correct book is found. This is done by using a while loop which doesn't stop until all the pages have been searched. Then a for loop is used to check all the books on the current page and a combination of if statements are used to check if any of the book matches the one being searched.

If book is found, then a message is logged on the console alongside the
details of the book. The details of the book are also returned, breaking
any unecessary for or while loops.

If book is not found, then a suitable message is logged on the console
and a null value is returned.