WebPage layout resembling Google with CSS
In this commit I have added a stylesheet file names style.css where I
have styled the existing index.html page to resemble the google website
when you are trying to search something.

The reason for this creative choice is because one of the tasks is to
search for a specific book so the user will more familiar with how to
search if the layour resembles Google. This is advantageous for user
experience.

Furthermore, to add the rest of the tasks; I wil ensure that if the
search bar is left empty then it will output all the books and the
checkboxes can be used to determine wether the subjects of the books
need to be mapped to uppercase and wether the filter needs to be applied
ton only show the books by author who lived in the last 200 years.

The next step is to code the script.js to ensure that there are no
logical errors in the output of this webpage.