Tasks can now be completed on webpage
With this commit, all the tasks that were previously done on the console
log can now be output on a webpage which is designed to resemble Google.

The only limitation with this commit is that when you click on a
specific book to access more information about it then it directs you to
a page where it shows the JSON containing all the information about the
book which is not necessarily easy to read.