Added a results counter to the search results
In this commit I have updated the index.html, styling.css and script.js
files to accomodate a new div for the result-counter.

This div will show a message once a search has been made, the message
will display 'Showing <number> results...' where number is the number of
results currently being shown on the web page.

The reason why I have added this feature is because popular search
engines such as Bing have this feature and that is what it is inspired
by. This feature will allow the user to also comprehend how many books
are being shown in the results and also calculate how the number of
results changes depending on which filters are applied.