Books in search result open book.html on click
In this commit I have created a new file called book.html. This is a
basic html file with an empty div where the information about the book
will be added through javascript.

Clicking on a book from the search result now opens this new file
book.html. To do this I added an onclick attribute to the text with the
book title. This onclick executes the function book_detail(id) with id
of the book as the parameter.

The function book_detail(id) uses this id to create a url which contains
a link to the book.html page with the id parameter embedded in the url.
In the next commit I will be fetching this parameter from the url and
processing it to output the information about the book related to this
id.