book.html loads the information about the  book
This commit adds a new javascript file called book.js which is executed
whenever book.html webpage is opened. This script uses the id parameter
in the url to fetch the information about the book from the
gutendex.com

This information is then added to a div in book.html through a
javascript function.

Next I need to make a new css file to style book.html