Styling for the book.html page
This commit adds class lists to the elements that are added to the
book.html through the book.js file. These classes are then used to style
the webpag using the already existing cascading stylesheet called
style.css.

I have decided to add a very generic styling to this page for now and
can come to it later if time permits but is unlikely. For now this
webpage will remain as it is.